Create Stop Position Markers on Edges
Overlaps and depth have been fixed. The size of the Stop Position Marker is determined using the mean depth. This means the marker might appear too small for edges close to the ego position but should roughly match the size of edges further away. Edges with lengths calculated as too small using the depth estimation are skipped. However, due to the use of rough depth estimation, some edges might be skipped unintentionally.